// components/log/log-list.interfaces.tsx
export interface IStudent {
  FIO: string;
  aliveType: string;
  data: string;
  dt_log: string;
  id: string;
  id_terminal: string;
  identifyType: string;
  local_ip_terminal: string;
  maskState: string;
  model: string;
  passTimeType: string;
  path: string;
  permissionTimeType: string;
  personId: string;
  qrCode: string;
  recModeType: string;
  recType: string;
  std_temperature: string;
  tempUnit: string;
  temperature: string;
  terminal_key: string;
  title_terminal: string;
  type: string;
  image: string;
}